refactor(frontend): moved landing page components to landing page dirctory
