+++ libs/ui-components/src/index.ts
@@ -4,17 +4,17 @@ import * as langSwticherI18n from './language-switcher/lang-switcher.i18n.json';
 import * as modelListI18n from './model/components/model-list/model-list.i18n.json';
 import * as modelUploadI18n from './model/components/model-upload/model-upload.i18n.json';
 import * as navBarI18n from './navbar/navbar.i18n.json';
-import * as stageI18n from './stage/stage.i18n.json';
-import * as toolbarI18n from './toolbar/toolbar.i18n.json';
+import * as stageI18n from './landing-page/stage/stage.i18n.json';
+import * as toolbarI18n from './landing-page/toolbar/toolbar.i18n.json';
 
 export * from './navbar/navbar.component';
 
-export * from './toolbar/toolbar.component';
+export * from './landing-page/toolbar/toolbar.component';
 export { toolbarI18n };
 
 export * from './model/components/model-list/model-list.component';
 export * from './model/components/model-upload/model-upload.component';
-export * from './stage/stage.component';
+export * from './landing-page/stage/stage.component';
 export * from './language-switcher/constants';
 export * from './language-switcher/language-switcher.component';
 export { stageI18n };
