+++ libs/ui-components/src/landing-page/stage/stage.component.scss
