+++ libs/ui-components/src/landing-page/stage/stage.component.spec.ts
