+++ libs/ui-components/src/landing-page/stage/stage.component.ts
