+++ libs/ui-components/src/landing-page/stage/stage.i18n.json
